+++ neo/renderer/DX12/dx12_RenderPass.cpp
@@ -74,7 +74,7 @@ void DX12Rendering::RenderPassBlock::UpdateRenderState(DX12Rendering::Commands::
 		std::vector<D3D12_RESOURCE_BARRIER> transitions;
 		transitions.reserve(renderTargetCount);
 
-		for (int index = 0; index < renderTargetCount; ++index)
+		for (UINT index = 0; index < renderTargetCount; ++index)
 		{
 			DX12Rendering::RenderSurface* outputSurface = DX12Rendering::GetSurface(surfaceList[index]);
 
@@ -87,7 +87,7 @@ void DX12Rendering::RenderPassBlock::UpdateRenderState(DX12Rendering::Commands::
 
 		if (transitions.size() > 0)
 		{
-			commandList->ResourceBarrier(transitions.size(), transitions.data());
+			commandList->ResourceBarrier(static_cast<UINT>(transitions.size()), transitions.data());
 		}
 	});
 }
