+++ neo/ui/Winvar.h
@@ -32,7 +32,7 @@ If you have questions concerning this license or the applicable additional terms
 #include "Rectangle.h"
 
 static const char *VAR_GUIPREFIX = "gui::";
-static const int VAR_GUIPREFIX_LEN = strlen(VAR_GUIPREFIX);
+static const int VAR_GUIPREFIX_LEN = static_cast<int>(strlen(VAR_GUIPREFIX));
 
 class idWindow;
 class idWinVar {
